+++ Web/App/controllers/account/loginController.js
@@ -12,36 +12,18 @@
         vm.getAccessToken = getAccessToken;
 
         function getAccessToken() {
-            userFactory.getCurrentUser()
-                .then(function (oldUser) {
+            userFactory.getAccessToken(vm.email, vm.password)
+                .success(function () {
 
-                    userFactory.getAccessToken(vm.email, vm.password)
-                        .success(function () {
-
-                            userFactory.getCurrentUser()
-                                .then(function (newUser) {
-
-                                    // Move anonymously created entities to this logged in user
-                                    userFactory.updateAnonymousChanges(oldUser, newUser)
-                                        .then(function () {
-
-                                            // Save changes
-                                            userFactory.saveChanges()
-                                                .then(function () {
-
-                                                    // Redirect the user to the previous page, except if it's login
-                                                    $location.path($rootScope.locationHistory[$rootScope.locationHistory.length - 2].path());
-                                                });
-                                        });
-                                });
-                        })
-                        .error(function (response) {
-                            if (typeof response.error_description !== 'undefined') {
-                                logger.logError(response.error_description, null, true);
-                            } else {
-                                logger.logError(response, null, true);
-                            }
-                        });
+                    // Redirect the user to the previous page
+                    $location.path($rootScope.locationHistory[$rootScope.locationHistory.length - 2].path());
+                })
+                .error(function (response) {
+                    if (typeof response.error_description !== 'undefined') {
+                        logger.logError(response.error_description, null, true);
+                    } else {
+                        logger.logError(response, null, true);
+                    }
                 });
         }
     };