+++ Web/App/controllers/account/registerController.js
@@ -38,25 +38,15 @@
         function register() {
             userFactory.register(vm)
                 .success(function () {
-
                     logger.logSuccess('You have been registered!', null, true);
-
-                    userFactory.getAccessToken(vm.email, vm.password)
-                        .success(function () {
-
-                            // Save changes
-                            userFactory.saveChanges()
-                                .then(function () {
-                                    $location.path('/account/confirmEmail');
-                                });
-                        })
-                        .error(function (response) {
-                            if (typeof response.error_description !== 'undefined') {
-                                logger.logError(response.error_description, null, true);
-                            } else {
-                                logger.logError(response, null, true);
-                            }
-                        });
+                    $location.path('/account/confirmEmail');
+                })
+                .error(function (response) {
+                    if (typeof response.error_description !== 'undefined') {
+                        logger.logError(response.error_description, null, true);
+                    } else {
+                        logger.logError(response, null, true);
+                    }
                 });
         }
     };