+++ Web/App/controllers/content/closingNotesController.js
@@ -12,17 +12,12 @@
 
         // View model
         var vm = this;
+        vm.isAuthenticated = false;
         vm.isLocalhost = $location.$$host === 'localhost';
         vm.displayBankTransfer = false;
-        vm.toggleBankTransfer = function () {
-            vm.displayBankTransfer = !vm.displayBankTransfer;
-        }
+        vm.toggleBankTransfer = toggleBankTransfer;
 
-        vm.isAuthenticated = false;
-        userFactory.isAuthenticated()
-            .then(function (isAuthenticated) {
-                vm.isAuthenticated = isAuthenticated;
-            });
+        _init();
 
         // User logged in & out
         $scope.$on('userLoggedIn', function () {
@@ -32,6 +27,17 @@
         $scope.$on('userLoggedOut', function () {
             vm.isAuthenticated = false;
         });
+
+        function _init() {
+            userFactory.getCurrentUser()
+                .then(function (currentUser) {
+                    vm.isAuthenticated = currentUser.isAuthenticated();
+                });
+        }
+
+        function toggleBankTransfer() {
+            vm.displayBankTransfer = !vm.displayBankTransfer;
+        }
 
         //vm.flattrIFrameUrl = $sce.trustAsResourceUrl('//api.flattr.com/button/view/?uid=forCrowd&button=compact&url=' + encodeURIComponent($location.$$absUrl));
     };