+++ Web/App_Start/MvcRouteConfig.cs
@@ -9,19 +9,10 @@
         {
             routes.IgnoreRoute("{resource}.axd/{*pathInfo}");
 
-            //routes.MapRoute(
-            //    name: "Error",
-            //    url: "Error/{action}");
-
-            //routes.MapRoute(
-            //    name: "Default",
-            //    url: "{*catchall}",
-            //    defaults: new { controller = "Home", action = "Index" });
-
             routes.MapPageRoute(
                 routeName: "Default",
                 routeUrl: "{*catchall}",
-                physicalFile: "~/main.aspx?v=0.41.2");
+                physicalFile: "~/main.aspx?v=0.41.3");
         }
     }
 }